Move SessionState to internal package
